Lógica terminada para el manejo desde el backend de una solicitud de elimación de empresa, manejo de errores y excepciones

+++ planilla_Backend/Planilla_Backend/CleanArchitecture/Application/UseCases/company/DeleteCompanyCommand.cs
@@ -13,8 +13,8 @@ namespace Planilla_Backend.CleanArchitecture.Application.UseCases.company
 
     public async Task<int> Execute(int companyUniqueId, int personId)
     {
-      if (personId < 0) throw new ArgumentException("Id de persona inválido");
-      if (companyUniqueId < 0) throw new ArgumentException("Id de la empresa inválido");
+      if (personId < 0) throw new ArgumentException("Id de persona inválido.");
+      if (companyUniqueId < 0) throw new ArgumentException("Id de la empresa inválido.");
 
       int isAdmin = await this.companyRepository.IsPersonAdmin(personId);
 
@@ -22,10 +22,16 @@ namespace Planilla_Backend.CleanArchitecture.Application.UseCases.company
       {
         int isOwner = await this.companyRepository.IsPersonOwnerOfCompany(companyUniqueId, personId);
 
-        if (isOwner != 1) throw new KeyNotFoundException("No tiene la autorización para realizar esta acción");
+        if (isOwner != 1) throw new KeyNotFoundException("No tiene la autorización para realizar esta acción.");
+
+        if (isOwner == -1) throw new Exception("Error al verificar permisos de la persona.");
       }
 
+      if (isAdmin == -1) throw new Exception("Error al verificar permisos de la persona.");
+
       int rowsAffected = await this.companyRepository.DeleteCompanyByUniqueId(companyUniqueId);
+
+      if (rowsAffected == -1) throw new Exception("Error al intentar eliminar la empresa.");
 
       return rowsAffected;
     }
